Remove url from track
Problem
-------

We moved to using filename in a previous PR.  As it turns out, the url
is no longer necessary.  Additionally, the move broke the app because we
were never saving filename for the new records (or with update).

Solution
--------

Update the track form to save filename instead of url and remove the url
from the model.